+++ autolamella/ui/AutoLamellaUI.py
@@ -19,6 +19,8 @@ from fibsem.structures import (
     FibsemRectangle,
     ImageSettings,
     MicroscopeSettings,
+    Point,
+    FibsemStagePosition,
 )
 from fibsem.ui.FibsemImageSettingsWidget import FibsemImageSettingsWidget
 from fibsem.ui.FibsemMovementWidget import FibsemMovementWidget
@@ -52,11 +54,15 @@ from autolamella.utils import INSTRUCTION_MESSAGES, check_loaded_protocol
 from PyQt5.QtCore import pyqtSignal
 
 from napari.qt.threading import thread_worker
+from fibsem.ui.FibsemMinimapWidget import FibsemMinimapWidget
+
 
 
 _DEV_MODE = True
 DEV_EXP_PATH = r"C:/Users/Admin/Github/autolamella/autolamella/log\ROHIT_TEST/experiment.yaml"
 DEV_PROTOCOL_PATH = cfg.PROTOCOL_PATH
+
+_AUTO_SYNC_MINIMAP = True
 
 def log_status_message(lamella: Lamella, step: str):
     logging.debug(f"STATUS | {lamella._petname} | {lamella.state.stage.name} | {step}")
@@ -114,7 +120,7 @@ class AutoLamellaUI(QtWidgets.QMainWindow, AutoLamellaUI.Ui_MainWindow):
 
 
     def setup_connections(self):
-        self.pushButton_add_lamella.clicked.connect(self.add_lamella_ui)
+        self.pushButton_add_lamella.clicked.connect(lambda: self.add_lamella_ui(pos=None))
         self.pushButton_add_lamella.setEnabled(False)
         self.pushButton_remove_lamella.clicked.connect(self.remove_lamella_ui)
         self.pushButton_remove_lamella.setEnabled(False)
@@ -139,6 +145,8 @@ class AutoLamellaUI(QtWidgets.QMainWindow, AutoLamellaUI.Ui_MainWindow):
         # self.actionSave_Protocol.triggered.connect(self.save_protocol)
         # self.actionEdit_Protocol.triggered.connect(self.edit_protocol)
         self.actionCryo_Sputter.triggered.connect(self._cryo_sputter)
+        self.actionLoad_Positions.triggered.connect(self._load_positions)
+        self.actionOpen_Minimap.triggered.connect(self._open_minimap)
 
 
         self.pushButton_yes.clicked.connect(self.push_interaction_button)
@@ -251,6 +259,55 @@ class AutoLamellaUI(QtWidgets.QMainWindow, AutoLamellaUI.Ui_MainWindow):
 
             self._microscope_ui_loaded = False
 
+    def _open_minimap(self):
+        if self.microscope is None:
+            napari.utils.notifications.show_warning(f"Please connect to a microscope first... [No Microscope Connected]")
+            return
+
+        if self.movement_widget is None:
+            napari.utils.notifications.show_warning(f"Please connect to a microscope first... [No Movement Widget]")
+            return
+
+
+        # TODO: should make this more generic i guess, but this is fine for now
+        self.viewer2 = napari.Viewer(ndisplay=2)
+        self.minimap_widget = FibsemMinimapWidget(self.microscope, self.settings, viewer=self.viewer2, parent=self)
+        self.viewer2.window.add_dock_widget(
+            self.minimap_widget, area="right", add_vertical_stretch=False, name="OpenFIBSEM Minimap"
+        )
+        self.minimap_widget._stage_position_moved.connect(self.movement_widget._stage_position_moved)
+        if _AUTO_SYNC_MINIMAP:
+            self.minimap_widget._stage_position_added.connect(self._update_stage_positions)
+        napari.run(max_loop_level=2)
+
+
+    def _update_stage_positions(self, position: FibsemStagePosition):
+        # add lamella to experiment from tile manager
+        if self.experiment is None:
+            logging.warning("No experiment loaded")
+            return
+
+        self.add_lamella_ui(position)
+
+    def _load_positions(self):
+        
+        
+        path = _get_file_ui( msg="Select a position file to load", 
+            path=self.experiment.path, 
+            _filter= "*yaml", 
+            parent=self)
+
+        if path == "":
+            napari.utils.notifications.show_info(f"No file selected..")
+            return
+
+        pdict = utils.load_yaml(path)
+        
+        positions = [FibsemStagePosition.__from_dict__(p) for p in pdict]
+
+        for pos in positions:
+            self.add_lamella_ui(pos)
+
     def update_ui(self):
         """Update the ui based on the current state of the application."""
 
@@ -265,6 +322,8 @@ class AutoLamellaUI(QtWidgets.QMainWindow, AutoLamellaUI.Ui_MainWindow):
         self.actionLoad_Protocol.setVisible(_experiment_loaded)
         # self.actionSave_Protocol.setVisible(_protocol_loaded)
         self.actionCryo_Sputter.setVisible(_protocol_loaded)
+
+        self.actionLoad_Positions.setVisible(_experiment_loaded and _microscope_connected)
 
         # workflow buttons
         # self.pushButton_setup_autoliftout.setEnabled(_microscope_connected and _protocol_loaded)
@@ -493,13 +552,16 @@ class AutoLamellaUI(QtWidgets.QMainWindow, AutoLamellaUI.Ui_MainWindow):
         logging.info(f"Moved to position of {lamella.info}.")
         self.movement_widget.update_ui_after_movement()
 
-    def add_lamella_ui(self):
+    def add_lamella_ui(self, pos:FibsemStagePosition=None):
 
         lamella = Lamella(
             path=self.experiment.path,
             _number=len(self.experiment.positions) + 1,
         )
         lamella.state.microscope_state = self.microscope.get_current_microscope_state()
+
+        if pos is not None:
+            lamella.state.microscope_state.absolute_position = deepcopy(pos)
 
         self.experiment.positions.append(deepcopy(lamella))
         
@@ -777,159 +839,6 @@ def mill_fiducial(
     finally:
         return lamella
 
-
-# def run_autolamella(
-#     microscope: FibsemMicroscope,
-#     experiment: Experiment,
-#     microscope_settings: MicroscopeSettings,
-#     image_settings: ImageSettings,
-#     current_alignment: bool,
-#     lamella_stages=list[FibsemMillingStage],
-# ):
-#     """
-#     Runs the AutoLamella protocol. This function iterates over the specified stages and Lamella positions in the `microscope_settings` protocol to mill a lamella for each position.
-
-#     Args:
-#         microscope (FibsemMicroscope): The FibsemMicroscope object representing the microscope to be used for milling.
-#         experiment (Experiment): The Experiment object representing the experiment where the lamella milling is taking place.
-#         microscope_settings (MicroscopeSettings): The MicroscopeSettings object containing the parameters for the microscope operation.
-#         image_settings (ImageSettings): The ImageSettings object containing the parameters for image acquisition.
-
-#     Returns:
-#         Experiment: The updated Experiment object after the successful milling of all the lamella positions specified in the `microscope_settings` protocol.
-#     """
-
-#     _microexpansion_used = any(
-#         [
-#             stage
-#             for stage in lamella_stages
-#             if stage.name == AutoLamellaWaffleStage.MicroExpansion.name
-#         ]
-#     )
-#     success = True
-#     lamella: Lamella
-#     for i, stage in enumerate(lamella_stages):
-#         curr_stage = AutoLamellaWaffleStage[stage.name]
-
-#         lamella: Lamella
-#         for j, lamella in enumerate(experiment.positions):
-#             _COMPLETE_STAGE = False
-#             if curr_stage is AutoLamellaWaffleStage.RoughCut:
-#                 if (
-#                     not _microexpansion_used
-#                 ):  # , check if last stage was fiducial milled
-#                     if lamella.state.stage is AutoLamellaWaffleStage.MillFeatures:
-#                         _COMPLETE_STAGE = True
-#                 elif lamella.state.stage is AutoLamellaWaffleStage.MicroExpansion:
-#                     _COMPLETE_STAGE = True
-#             elif lamella.state.stage.value == curr_stage.value - 1:
-#                 _COMPLETE_STAGE = True
-
-#             if not _COMPLETE_STAGE:
-#                 continue
-
-#             lamella.state.start_timestamp = datetime.timestamp(datetime.now())
-#             log_status_message(lamella, "MOVING_TO_POSITION")
-#             microscope.move_stage_absolute(
-#                 lamella.state.microscope_state.absolute_position
-#             )
-#             log_status_message(lamella, "MOVE_TO_POSITION_SUCCESSFUL")
-
-#             image_settings.save_path = lamella.path
-#             image_settings.save = True
-#             image_settings.label = f"start_mill_stage_{i}"
-#             image_settings.reduced_area = None
-#             acquire.take_reference_images(microscope, image_settings)
-#             image_settings.save = False
-
-#             # alignment
-#             for _ in range(
-#                 int(microscope_settings.protocol["lamella"]["beam_shift_attempts"])
-#             ):
-#                 log_status_message(lamella, "BEAM_ALIGNMENT")
-#                 if current_alignment:
-#                     if isinstance(microscope, ThermoMicroscope) or isinstance(
-#                         microscope, DemoMicroscope
-#                     ):
-#                         microscope.set(
-#                             "current", stage.milling.milling_current, BeamType.ION
-#                         )
-#                     elif isinstance(microscope, TescanMicroscope):
-#                         microscope.set("preset", stage.milling.preset, BeamType.ION)
-#                 image_settings.beam_type = BeamType.ION
-#                 image_settings.reduced_area = lamella.fiducial_area
-#                 beam_shift_alignment(
-#                     microscope=microscope,
-#                     image_settings=image_settings,
-#                     ref_image=lamella.reference_image,
-#                     reduced_area=lamella.fiducial_area,
-#                 )
-
-#             try:
-#                 stage.milling.hfw = lamella.state.microscope_state.ib_settings.hfw
-#                 log_status_message(lamella, f"MILLING_TRENCH")
-
-#                 milling.setup_milling(
-#                     microscope,
-#                     mill_settings=stage.milling,
-#                 )
-
-#                 # redefine pattern for each lamella
-#                 stage.pattern.define(stage.pattern.protocol, lamella.lamella_centre)
-
-#                 milling.draw_patterns(
-#                     microscope=microscope,
-#                     patterns=stage.pattern.patterns,
-#                 )
-
-#                 milling.run_milling(
-#                     microscope, milling_current=stage.milling.milling_current
-#                 )
-#                 milling.finish_milling(microscope)
-#                 lamella.state.end_timestamp = datetime.timestamp(datetime.now())
-#                 image_settings.save_path = lamella.path
-#                 image_settings.reduced_area = None
-#                 log_status_message(lamella, f"MILLING_COMPLETED_SUCCESSFULLY")
-
-#                 # Update Lamella Stage and Experiment
-#                 lamella = lamella.update(stage=curr_stage)
-
-#                 # save reference images
-#                 image_settings.save = True
-#                 image_settings.label = f"ref_mill_stage_{i}"
-#                 image_settings.reduced_area = None
-#                 acquire.take_reference_images(microscope, image_settings)
-
-#                 image_settings.save = False
-
-#                 experiment.save()
-
-#                 l_stage = stage.name
-
-#                 logging.info(f"Lamella {j+1}, stage: '{l_stage}' milled successfully.")
-#                 log_status_message(lamella, f"STAGE_COMPLETE")
-
-#                 success = True
-#             except Exception as e:
-#                 logging.error(
-#                     f"Unable to draw/mill the lamella: {traceback.format_exc()}"
-#                 )
-#                 lamella.state.stage = AutoLamellaWaffleStage.MillFeatures
-#                 success = False
-#             finally:
-#                 milling.finish_milling(microscope)
-
-#     if success:
-#         logging.info("All Lamella milled successfully.")
-#     else:
-#         logging.info("Lamellas were not milled successfully.")
-#     for lamella in experiment.positions:
-#         if lamella.state.stage == AutoLamellaWaffleStage.PolishingCut:
-#             lamella = lamella.update(stage=AutoLamellaWaffleStage.Finished)
-
-#     experiment.save()
-
-#     return experiment
 
 
 def splutter_platinum(microscope: FibsemMicroscope):